refactor: メニューのアイテムの表示状態のコントロールを別クラスで行うようにしました
