+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/activity/UserDetailActivity.kt
@@ -47,6 +47,7 @@ import net.pantasystem.milktea.note.view.ActionNoteHandler
 import net.pantasystem.milktea.note.viewmodel.NotesViewModel
 import net.pantasystem.milktea.user.PinNoteFragment
 import net.pantasystem.milktea.user.R
+import net.pantasystem.milktea.user.activity.binder.UserDetailActivityMenuBinder
 import net.pantasystem.milktea.user.databinding.ActivityUserDetailBinding
 import net.pantasystem.milktea.user.nickname.EditNicknameDialog
 import net.pantasystem.milktea.user.viewmodel.UserDetailViewModel
@@ -316,42 +317,8 @@ class UserDetailActivity : AppCompatActivity() {
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         menuInflater.inflate(R.menu.activity_user_menu, menu)
 
-        val state = mViewModel.userState.value
-        Log.d("UserDetailActivity", "onCreateOptionsMenu: state:$state")
-
-        val block = menu.findItem(R.id.block)
-        val mute = menu.findItem(R.id.mute)
-        val unblock = menu.findItem(R.id.unblock)
-        val unmute = menu.findItem(R.id.unmute)
-        val report = menu.findItem(R.id.report_user)
-        mute?.isVisible = !(state?.isMuting ?: false)
-        block?.isVisible = !(state?.isBlocking ?: false)
-        unblock?.isVisible = (state?.isBlocking ?: false)
-        unmute?.isVisible = (state?.isMuting ?: false)
-        if (mViewModel.isMine.value) {
-            block?.isVisible = false
-            mute?.isVisible = false
-            unblock?.isVisible = false
-            unmute?.isVisible = false
-            report?.isVisible = false
-        }
-
-        val tab = menu.findItem(R.id.nav_add_to_tab)
-        val page = accountStore.currentAccount?.pages?.firstOrNull {
-            val pageable = it.pageable()
-            if (pageable is Pageable.UserTimeline) {
-                pageable.userId == mUserId?.id
-            } else {
-                false
-            }
-        }
-        if (page == null) {
-            tab?.setIcon(R.drawable.ic_add_to_tab_24px)
-        } else {
-            tab?.setIcon(R.drawable.ic_remove_to_tab_24px)
-        }
-
-        applyMenuTint(this, menu)
+        UserDetailActivityMenuBinder(this, mViewModel, applyMenuTint, accountStore)
+            .bind(menu)
 
         return super.onCreateOptionsMenu(menu)
     }
